proto_tech3-main/requirements.txt
# cairo rendering support
cairocffi==0.6
fastapi==0.115.2
uvicorn==0.32.0
python-multipart==0.0.20
requests==2.32.3
rarfile==4.2
gspread==6.2.1
oauth2client==4.1.3
sqlalchemy==2.0.43

gunicorn==22.0.0           # Production-grade web server (WSGI)
alembic==1.13.1            # For database migrations
psycopg2-binary==2.9.9     # PostgreSQL database driver
PyJWT-2.10.1
# === Authentication & Security ===
passlib[bcrypt]==1.7.4     # For password hashing (includes bcrypt)
python-jose[cryptography]==3.3.0 # A robust library for JWT operations, often used by FastAPI's security utils
authlib==1.3.0             # For Google OAuth 2.0 integration

# === API Utilities ===
python-dotenv==1.0.1       # For loading .env files
slowapi==0.1.9             # For rate limiting
pydantic-settings==2.2.1   # For loading settings from .env into Pydantic models

# === Email Services ===
fastapi-mail==1.4.1        # For sending emails from the auth module (verification, etc.)
email-validator==2.1.1     # For validating email addresses in Pydantic models

# Stripe Integration
stripe==9.9.0              # The official Stripe Python library

# Other HTTP clients
httpx==0.27.0              # Modern async HTTP client (used by TestClient and Authlib)
requests==2.32.3           # For the EGP/YUAN currency conversion

rarfile==4.2               # For extracting .rar archives

# 3D Printing / Google Sheets
google-auth-oauthlib==1.2.0  # Authentication handler for gspread
pandas==2.2.2              # For data manipulation from Google Sheets

# === Redis ===
redis==5.0.4               # For connecting to Redis (used by the token denylist service)
itsdangerous==2.2.0

# Python 3.10.12
# sudo apt-get install unrar